Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>
Implemented most of the EPOS2 class functionality.
Individual motor variables are contained within a struct (eposMotor).
We keep a QList of shared pointers to these structs, so that we can
easily iterate over the list.

Clicking on Open in the EPOS2 GUI tries to connect to all nodes. If not
successful, the m_motorsEnabled state will be false. However if any of
the motors were successfuly enabled, those individual motors will have
their m_enabled state set to true, and these motors can still be manually
commanded. This is for testing purposes and might not make sense in final
deployment.

In manual control mode, a drop-down menu lets the user select which
node (motor) to control. This motor can be then sent to a desired QC
in absolute of relative counts. The motors can be homed collectively.
The motor motion can be halted.

The EPOS2 widget was added to the main window. This is not the final
positioning, a tab widget will be used to switch between different tabs.