Changed the structure of the LabJack worker-thread layout. Now the worker has an internal QTimer that calls the readData function every 1000 ms.
Added button functionality to connect to, initialize, start/stop recording
from, and disconnect from the Labjack. Disconnect needs further testing,
however the other functions seem to be working so far.

The main GUI will have to set the output filename and the epoch, which
is the QTime when the GUI was started. So mainwindow should
have a QTime that is .start()'ed during initialization.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>