Merged the widgets with the main GUI.
We are able to log EM data at the moment! The data is being passed by value at the moment, but we can use a shared_pointer to pass it also. The thread classes have to be public members in the widget classes for the GUI to be able to connect those signals. We should save the EM data in binary format, text output will take up too much space.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>