DataLoggerWidget:
Now each log file initialization creates a folder using the current
date and time. QDir is used to check folder existence and creation.

Switched to read only mode instead of disabled for line edits.

Now reporting number of entries logged for each file when files are closed.
Except when program is exit before close files button is clicked, in which
case this info doesn't get printed since the button was not pressed.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>