Ascension: The rotations are coming in inverted. Now transposing (= inverting) the rotation matrix reported by the trackers before emitting signals to the rest of the GUI.
Controller:
The paramteters used to initialize the Kinematics was in meters.
Now it's in mm.

Kinematics:
The singular configuration (when alpha = 0) was checked using a strictly
not equal to, but the variable is a double and we can have numerical
instabilities when this value is too small. Now the condition is
whether alpha is within epsilon 1e-6.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>