Controller: Moved cycle_recalculate to the new CyclicModel class. Changed the limits of the task space x and y axes from +/-100 to +/-50.
CyclicModel:
New class for creating respiration models (although, the class is
generic, and could be used to create a model for cardiac motion as well.)

It takes in readings from the EM sensors and stores it in a deque.
Deque is like a vector, but it's O(1) at pop/push_front/back, as well
as random element access.
The deque has a fixed size, and adding observation beyond this size
causes the oldest reading to get popped.

Filtfilt:
Found a C++ implementation of MATLAB's filtfilt function on the internet.
Will use for low pass filtering. Filtfilt may not be the best choice
for real-time applications though, since it's non-causal. An FIR/IIR
filter would be better, where we can choose an odd window size and easily
account for the phase shift.

Kinematics_4DOF:
fZeroAlpha will return 0.0 if yprime is very small or negative.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>