CyclicModel: Added functions to load data from previous experiments, and to test the low pass filter.
filtfilt:
Optimized filtering so that most of the computation is done when
filter parameters are set. For 4000 data points, filtering takes
about 800 microseconds. This could still be improved, perhaps using
Eigen::Vectors.
But the 4000 data points is only going to be filtered at initialization
of the breathing model. Past that, filtering 51(50?) data points, where
the last element is the latest reading, should be fine.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>