filtfilt: Implemented simultaneous multi-axis filtering, using Eigen::Vector types stored in a deque (vector implementation might be faster?).
CycleModel:
Implemented data transfer from Transform to each axis, and connected
to filtfilt.
Added test code for multi-axis data loading and filtering.
Plotting the results in MATLAB, I see that the edge effects are pretty bad.
There is a bug somewhere that is only affecting the edges, probably a
problem with the padding on the sides.

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>