respModelWidget: Added a new Widget to take care of the respiration model user interaction.
Need to add communication between controller thread and this widget to
keep track of initialization and state.

It would be nice to be able to change things like heart rate, breath rate,
sample size, etc. from the GUI, but not a high priority.

cyclicModel:
Deleted some of the commented out code to make things neater. Especially
the std::vector<std::vector<double>> code was deleted.

Cycle_raclaculate is overloaded to take both 7d and 1d data. Will
see if running all 7 at once is faster than 1 at a time.

###
Need to finalize the implementation of the period update details.
###

Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>