There was a performance bottleneck with the cycle_recalculate. For 4000 points, it took 10-20 ms per cycle_recalc, so the control cycle took 40-60ms!!! This caused EM readings to pile up ### TODO: (it might be a good idea to add a perish time for these emitted readings - if they're not read within 7ms, delete?)
Now each of the threee cycle_recalculates run on a separate thread
using QtConcurrent. QFuture grabs the results when they become available.
It's OK for the cycle_recalc not to run every cycle, the model should
be accurate for a few cycles. [By the way, QtConcurrent doesn't like
overloaded functions, so I gave the matrix function and the vector
function different names *M and *V.]

Now the control cycles take about 2ms, and the robot is responsive.


Implemented some of the RespComp GUI elements, but needs refinement.
Plotting would be good, but need to look into the computational cost.


Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>