Implemented new Jacobian-based controller. Maps task to config space. Behaving well! Uses the global heading I derived. When driving the robot using the task space though, this has the disadvantage that going from (x,y) = (4,0) to (0,4) with constant psi will cause a 90 degree roll to the cathteter. Does it make sense to redefine psi as roll? Probably not. Then there should be a mode where psi corresponds to roll just for driving around.
Noticed that when the cath is at (x,y) = (0,10), it's having trouble
converging. May be an issue with this catheter, or an issue with
too much disturbance at the base, since the cath body is essentially
lifting up (kind of like an iverted pendulum problem!). Jacobian learning
may be beneficial to resolving this.


Signed-off-by: Degirmenci, Alperen <adegirmenci@seas.harvard.edu>