Cyclic Model
Trying to identify why the frequency is getting higher. This in turn
causes the program to crash.

Placed period update calls right after qConcurrent returns, and right
before another batch of qConcurrents are launched. This way, the new
calculations use the new period.

Noticed I was doing a max instead of a min in the peakDetector.

Using a NonRecursive mutex now.