Fix CyclicModel Eigen assert fail
Eigen::VectorXd timeDiff was initialized to have size mean.size(), but
should have been m_respPeakMean.size() instead. This was causing access
outside data bounds, leading to an assert failure in Eigen. I was able
to identify this issue by compiling and running (in Debug mode) the Qt
project in Visual Studio. The Qt add-on  for VS is really useful.