Sweep functionality
Created a Sweep class to keep track of sweep settings. Controller always
checks if there is a sweep happening. Based on the current psi, Sweep
determines if convergence is achieved, and initiates a timer to send the
next sweep command. TODO: Convergence should initiate image transmission
in FrameClient. This will probably require the Sweep to return an int
rather than a bool.

FrameClient is set to transmit images that are only in the first 6% of
the cardiac phase. This is because the Reconstruction code doesn't do 4D
yet, only 3D.

Working on 4D reconstruction in MATLAB. Copied code from the MATLAB GUI,
still in progress.