Respiration model and compensation working!
No longer recalculating the Fourier serier parameters for the CT or BB.
For predictions, the phase is the current phase of the Bird4 model is
used to project the CT/BB model forward.

Using an averaging filter for the period update. This makes the
transition smoother.

Now able to change the focus of the plot in the resp model widget,
between Bird4 and CT x-axis.